+++ Apps/Examples/Examples/All Examples/ViewportExample.swift
@@ -36,11 +36,11 @@ final class ViewportExample: UIViewController, ExampleProtocol {
         mapView.mapboxMap.setCamera(to: CameraOptions(center: cupertino, zoom: 14))
 
         mapView.location.options.puckType = .puck2D(.makeDefault(showBearing: true))
-        mapView.location.options.puckBearingSource = .course
+        mapView.location.options.puckBearingSource = .heading
 
         followPuckViewportState = mapView.viewport.makeFollowPuckViewportState(
             options: FollowPuckViewportStateOptions(
-                bearing: .course))
+                bearing: .heading))
 
         overviewViewportState = mapView.viewport.makeOverviewViewportState(
             options: OverviewViewportStateOptions(