+++ Sources/MapboxMaps/Foundation/CameraBounds.swift
@@ -4,20 +4,25 @@ import CoreLocation
 /// Holds information about camera bounds.
 public struct CameraBounds: Hashable {
     /// The latitude and longitude bounds to which the camera center are constrained.
+    /// Defaults to: Southwest: (-90, -180) and Northeast: (90, 180).
     public let bounds: CoordinateBounds
 
     /// The maximum zoom level, in mapbox zoom levels 0-25.5. At low zoom levels,
     /// a small set of map tiles covers a large geographical area. At higher zoom
     /// levels, a larger number of tiles cover a smaller geographical area.
+    /// Defaults to 22.
     public let maxZoom: CGFloat
 
     /// The minimum zoom level, in mapbox zoom levels 0-25.5.
+    /// Defaults to 0.
     public let minZoom: CGFloat
 
     /// The maximum allowed pitch value in degrees.
+    /// Defaults to 85.
     public let maxPitch: CGFloat
 
     /// The minimum allowed pitch value degrees.
+    /// Defaults to 0.
     public let minPitch: CGFloat
 
     internal init(bounds: CoordinateBounds,