+++ Sources/MapboxMaps/Foundation/MapInitOptions.swift
@@ -16,10 +16,10 @@ public final class MapInitOptions: NSObject {
     /// Associated `MapOptions`
     public let mapOptions: MapOptions
 
-    /// Default style URI for initializing the map
+    /// Style URI for initializing the map. Defaults to Mapbox Streets.
     public let styleURI: StyleURI?
 
-    /// Default camera options for initializing the map
+    /// Camera options for initializing the map. CameraOptions default to 0.0 for each value.
     public let cameraOptions: CameraOptions?
 
     /// Initializer. The default initializer, i.e. `MapInitOptions()` will use